Created more front end pages(HTML Pages for Initiatives)
I am hoping that I committed the correct files and don't mess everything
up.

I added the controller and view for the 2 Initiative pages, currently
they are just links that go nowhere(but its a start). At this point I am
unsure about how the controller and view will work. I think there needs
to be a controller for each Initiative but only 1 view that they al
share(Again I could be wrong)

When you run the app select InitiativeController to see this first page
and InitiativePageController to see the second.

Let me know if you have problems running the application and I will try
to recommit it.